+++ src/translations/en_US.js
@@ -14,6 +14,7 @@ module.exports = {
   genericDeleteConfirmationMessage: 'Are you sure you want to delete this payment method?',
   deleteCancelButton: 'Cancel',
   deleteConfirmationButton: 'Delete',
+  hasSubscription: 'This payment method cannot be deleted because it is associated with a recurring payment.',
   cardVerification: 'Card Verification',
   // Errors
   fieldEmptyForCvv: 'Please fill out a CVV.',